Make s3 client call function fix minor issues and change deploy workflow ready for testing it out

+++ videos/tasks.py
@@ -6,6 +6,8 @@ from django.http import JsonResponse
 from django.conf import settings
 from celery import shared_task
 
+from videos.utils import get_s3_client
+
 import logging
 
 from videos.models import Video
@@ -16,12 +18,10 @@ logger = logging.getLogger("django_mcq")
 @shared_task
 def delete_s3_file(video_id: int):
 
-    s3 = boto3.client(
-        's3',
-        aws_access_key_id=settings.AWS_ACCESS_KEY,
-        aws_secret_access_key=settings.AWS_SECRET_ACCESS_KEY,
-        region_name=settings.AWS_REGION
-    )
+    s3 = get_s3_client()
+
+    if s3 is None:
+        raise Exception("S3 Client not initialized")
 
     try:
         s3.delete_object(Bucket=settings.S3_BUCKET_NAME, Key=f"videos/{video_id}.mp4")
